pipeline/requirements.txt
pyarrow
pandas
supabase
python-dotenv
